requirements.txt
pyyaml
numpy 
pytorch-lightning
einops
scikit-learn
pandas
opencv-python
nibabel